+++ CHANGES.md
@@ -1,6 +1,7 @@
 0.1.0
 -----
 * Factor 0MQ bindings into the celluloid-zmq gem
+* Heartbeat system for detecting downed nodes
 
 0.0.1
 -----